+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v?.?.?
+?/?/?
+- Added code comments
+
 # v0.4
 2015/7/5
 - Added option for custom log message sink and maximum log message size